feat: allow multiple product selection with hint
Add multiproduct variant for product selection screen that lets users
build a stack of products. Refactor product selection into variant
directories and add shared ProductCarousel component.
