+++ frontend/src/scenes/onboarding/productSelection/productSelectionLogic.ts
@@ -20,7 +20,7 @@ import {
 import type { productSelectionLogicType } from './productSelectionLogicType'
 
 export type OnboardingStep = 'choose_path' | 'product_selection'
-export type RecommendationSource = 'use_case' | 'browsing_history' | 'manual' | 'simplified'
+export type RecommendationSource = 'use_case' | 'browsing_history' | 'manual' | 'simplified' | 'multiproduct'
 
 export const productSelectionLogic = kea<productSelectionLogicType>([
     path(['scenes', 'onboarding', 'productSelection', 'productSelectionLogic']),
