+++ frontend/src/scenes/onboarding/productSelection/variants/spotlight/SpotlightProductSelection.scss
