+++ frontend/src/scenes/onboarding/productSelection/variants/spotlight/SpotlightProductSelection.tsx
@@ -1,0 +1,5 @@
+import { ProductCarousel } from '../ProductCarousel'
+
+export function SpotlightProductSelection(): JSX.Element {
+    return <ProductCarousel mode="single" recommendationSource="simplified" />
+}
